# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - OpenAPI documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/brands": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "List brands",
                "operationId": "dfb301a82846380eaa3335b289acea8e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results in page. default is 10.",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "The field to which the results are ordered by. default is Id.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The limit of results returned. returns all matching resutls if no limit is provided.",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Specifies whether to order the results descendingly or ascendingly. default is false. (asc)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/brand/{uuid}": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Fetch a single brand",
                "operationId": "bb44258513559db36dbf030a5669ffb5",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succses"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Brand"
                ],
                "summary": "Update a brand",
                "operationId": "5838f0ec90ab49176e59fe9dbcb4f24b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The unique UUID of the brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "The title of the brand",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succses"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Brand"
                ],
                "summary": "Update a brand",
                "operationId": "f8b824c5142920ff178e441e70ca443e",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The unique UUID of the brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succses"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/brand/create": {
            "post": {
                "tags": [
                    "Brand"
                ],
                "summary": "Create a new brand",
                "operationId": "2fb74671e7df923dca403352b662464b",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "The title of the brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succses"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}